Refactor FrozenTreasure to be in events.jerry.frozentreasure
